+++ Model/Reepayapplepay.php
@@ -1,0 +1,13 @@
+<?php
+ 
+namespace Radarsofthouse\Reepay\Model;
+
+/**
+ * Class Reepayapplepay
+ *
+ * @package Radarsofthouse\Reepay\Model
+ */
+class Reepayapplepay extends \Radarsofthouse\Reepay\Model\Reepaypayment
+{
+    protected $_code = 'reepay_applepay';
+}