+++ Model/Reepayforbrugsforeningen.php
@@ -1,0 +1,13 @@
+<?php
+ 
+namespace Radarsofthouse\Reepay\Model;
+
+/**
+ * Class Reepayforbrugsforeningen
+ *
+ * @package Radarsofthouse\Reepay\Model
+ */
+class Reepayforbrugsforeningen extends \Radarsofthouse\Reepay\Model\Reepaypayment
+{
+    protected $_code = 'reepay_forbrugsforeningen';
+}