+++ Model/Reepayklarnapaylater.php
@@ -1,0 +1,13 @@
+<?php
+ 
+namespace Radarsofthouse\Reepay\Model;
+
+/**
+ * Class Reepayklarnapaylater
+ *
+ * @package Radarsofthouse\Reepay\Model
+ */
+class Reepayklarnapaylater extends \Radarsofthouse\Reepay\Model\Reepaypayment
+{
+    protected $_code = 'reepay_klarnapaylater';
+}