+++ Model/Reepayklarnapaynow.php
@@ -1,0 +1,13 @@
+<?php
+ 
+namespace Radarsofthouse\Reepay\Model;
+
+/**
+ * Class Reepayklarnapaynow
+ *
+ * @package Radarsofthouse\Reepay\Model
+ */
+class Reepayklarnapaynow extends \Radarsofthouse\Reepay\Model\Reepaypayment
+{
+    protected $_code = 'reepay_klarnapaynow';
+}