+++ Model/Reepaypaypal.php
@@ -1,0 +1,13 @@
+<?php
+ 
+namespace Radarsofthouse\Reepay\Model;
+
+/**
+ * Class Reepaypaypal
+ *
+ * @package Radarsofthouse\Reepay\Model
+ */
+class Reepaypaypal extends \Radarsofthouse\Reepay\Model\Reepaypayment
+{
+    protected $_code = 'reepay_paypal';
+}