+++ Model/Reepayresurs.php
@@ -1,0 +1,13 @@
+<?php
+ 
+namespace Radarsofthouse\Reepay\Model;
+
+/**
+ * Class Reepayresurs
+ *
+ * @package Radarsofthouse\Reepay\Model
+ */
+class Reepayresurs extends \Radarsofthouse\Reepay\Model\Reepaypayment
+{
+    protected $_code = 'reepay_resurs';
+}