+++ Model/Reepayswish.php
@@ -1,0 +1,13 @@
+<?php
+ 
+namespace Radarsofthouse\Reepay\Model;
+
+/**
+ * Class Reepayswish
+ *
+ * @package Radarsofthouse\Reepay\Model
+ */
+class Reepayswish extends \Radarsofthouse\Reepay\Model\Reepaypayment
+{
+    protected $_code = 'reepay_swish';
+}